cl-gopher.lisp: Make uri printing and parsing symmetric.
This changes uri-for-gopher-line in a backwards-compatible manner, to
be more consistent with parse-gopher-uri's behavior. To be more
precise, uri-for-gopher-line prints no port it it's 70, much like
parse-gopher-uri using port 70 if it's not provided.

In other words,
(string= (uri-for-gopher-line (parse-gopher-uri uri)) uri)
should be true for all the valid Gopher URIs.

As an unnecessary stylistic change, conditional formatters are used
instead of explicit if-s to reduce code duplication.